fix segfault on unloading textures with misconfigured scene
