+++ src/init/get_extras.c
@@ -12,26 +12,21 @@
 
 #include "cub3D.h"
 
-int	sprites_to_array(t_scene *scn)
+inline static int	check_file(char const *path)
 {
-	t_size		i;
-	t_sprite	*spr;
+	int	fd;
 
-	scn->spr_arr = malloc(sizeof(t_sprite *) * scn->n_sprites);
-	if (!scn->spr_arr)
-		return (throw_error("Memory allocation failed"));
-	spr = scn->sprites;
-	i = 0;
-	while (i < scn->n_sprites && spr)
-	{
-		scn->spr_arr[i++] = spr;
-		spr = spr->next;
-	}
+	fd = open(path, O_RDONLY);
+	if (fd < 0)
+		return (throw_error("Can't open texture file"));
+	close(fd);
 	return (0);
 }
 
 static void	set_sprite_params(t_sprite *spr, char const **param)
 {
+	spr->next = NULL;
+	spr->isloaded = 0;
 	spr->position.x = ft_atod_dirty(param[1]);
 	spr->position.y = ft_atod_dirty(param[2]);
 	spr->hdiv = 1;
@@ -56,10 +51,11 @@ int	get_sprite(t_scene *scene, char const **param)
 
 	if (!param[1] || !param[2] || !param[3])
 		return (throw_error("Wrong number of parameters"));
+	if (check_file(param[3]))
+		return (1);
 	spr = malloc(sizeof(t_sprite));
 	if (!spr)
 		return (throw_error("Memory allocation failure"));
-	spr->next = NULL;
 	spr->texture = ft_strdup(param[3]);
 	if (!spr->texture && (free(spr), 1))
 		return (throw_error("Memory allocation failure"));
@@ -77,6 +73,22 @@ int	get_sprite(t_scene *scene, char const **param)
 	return (0);
 }
 
+static int	set_extra_params(t_extra *ex, char const **param)
+{
+	ex->key = ft_strdup(param[0]);
+	if (!ex->key)
+		return (throw_error("Memory allocation failure"));
+	ex->value = ft_strdup(param[1]);
+	if (!ex->value)
+	{
+		free(ex->key);
+		return (throw_error("Memory allocation failure"));
+	}
+	ex->next = NULL;
+	ex->isloaded = 0;
+	return (0);
+}
+
 int	get_extra(t_scene *scene, char const **param)
 {
 	t_extra	*ex;
@@ -86,14 +98,13 @@ int	get_extra(t_scene *scene, char const **param)
 		return (get_sprite(scene, param));
 	if (!param[1] || param[2])
 		return (throw_error("Wrong number of parameters"));
+	if (check_file(param[1]))
+		return (1);
 	ex = malloc(sizeof(t_extra));
 	if (!ex)
 		return (throw_error("Memory allocation failure"));
-	ex->key = ft_strdup(param[0]);
-	ex->value = ft_strdup(param[1]);
-	if (!ex->key || (!ex->value && (free(ex->key), 1)))
-		return (throw_error("Memory allocation failure"));
-	ex->next = NULL;
+	if (set_extra_params(ex, param))
+		return (free(ex), 1);
 	if (!scene->extras)
 		scene->extras = ex;
 	else
